Move bridge to separate file
